+++ src/main/java/org/opentripplanner/graph_builder/module/osm/WayPropertySetSource.java
@@ -21,6 +21,8 @@ public interface WayPropertySetSource {
 			return new DefaultWayPropertySetSource();
 		} else if ("norway".equals(type)) {
 			return new NorwayWayPropertySetSource();
+		} else if ("uk".equals(type)) {
+			return new UKWayPropertySetSource();
 		} else {
 			throw new IllegalArgumentException(String.format("Unknown osmWayPropertySet: '%s'", type));
 		}